migrated new plan component to create plan
